Remove PDF encryption support
