+++ README.md
@@ -12,7 +12,6 @@ This web application manages the issuance and storage of course certificates. It
    ```
 2. **Configure environment variables** – copy `.example.env` to `.env` and update the values to match your setup. Provide the credentials for your external PostgreSQL server by setting `POSTGRES_HOST`, `POSTGRES_DB`, `POSTGRES_USER`, `POSTGRES_PASSWORD`, and optionally `POSTGRES_PORT`. The container derives the SQLAlchemy connection string from these values whenever `DATABASE_URL` is empty. If you prefer, you can instead set `DATABASE_URL` directly to a PostgreSQL connection string. For local development or automated tests you may temporarily enable `ENABLE_LOCAL_TEST_DB` to let the app create a SQLite database file defined by `LOCAL_TEST_DB_PATH`; the flag is disabled by default so production deployments still require PostgreSQL.
 
-   *Säker PDF-lagring*: PDF-intyg krypteras med AES-GCM med en nyckel som härleds via samma PBKDF2-baserade hashning och `HASH_SALT` som används för lösenord och personnummer. Se därför till att `HASH_SALT` är satt i `.env` och håll värdet stabilt över tid så att uppladdade filer kan dekrypteras.
 3. **Run the application**
    ```bash
    python app.py
