+++ SECURITY.md
@@ -24,6 +24,3 @@ Inte delar informationen offentligt förrän vi har hunnit åtgärda problemet.
 
 Tillsammans kan vi skapa en tryggare digital miljö.
 
-## Kryptering av lagrade PDF:er
-
-Alla PDF-intyg krypteras nu med AES-GCM innan de sparas i databasen. Krypteringsnyckeln härleds via samma PBKDF2-baserade hashfunktion och `HASH_SALT` som används för att skydda lösenord, personnummer och e-postadresser. Säkerställ därför att `HASH_SALT` är satt i din `.env`-fil och håll värdet oförändrat så länge du behöver komma åt befintliga intyg.
