+++ requirements.txt
@@ -6,7 +6,6 @@ itsdangerous==2.1.2
 click==8.1.7
 SQLAlchemy==2.0.43
 psycopg[binary]==3.2.10
-cryptography==46.0.1
 
 # Optional (for PDF validation if you later add more advanced checks)
 # PyPDF2==3.0.0
